Rename file probabilities.hpp to joint_probabilities.hpp.
